modified getThis() to return intptr_t instead of long to work with mingw64; added note about eigen3 in README.md; disabled CsoundAC for interim until allegro.h can be fixed for intptr_t

+++ frontends/CsoundAC/ScoreModel.hpp
@@ -67,7 +67,7 @@ public:
         /**
          * Returns the address of this as a long integer.
          */
-        virtual long getThis();
+        virtual intptr_t getThis();
         /**
          * Returns the address of this as a Node pointer.
          */
